refactor: move InMemorySecureKeyProvider to test project
Remove InMemorySecureKeyProvider from production code to prevent silent
key loss on app restart. Platforms must now register a real ISecureKeyProvider
or app startup will fail with PlatformNotSupportedException.
