+++ src/sdk/Angor.Sdk/Wallet/WalletContextServices.cs
@@ -37,7 +37,8 @@ public static class WalletContextServices
         services.AddSingleton<IWalletFactory, WalletFactory>();
         services.AddSingleton<IWalletOperations, WalletOperations>();
         services.AddSingleton<IPsbtOperations, PsbtOperations>();
-        services.TryAddSingleton<ISecureKeyProvider, InMemorySecureKeyProvider>();
+        // ISecureKeyProvider must be registered by the platform (DPAPI, Android KeyStore, Linux, etc.)
+        // before calling Register(). No fallback is provided — missing registration will fail at resolve time.
         services.TryAddSingleton<ISensitiveWalletDataProvider, SensitiveWalletDataProvider>();
         services.TryAddSingleton<IPasswordProvider, SecureKeyPasswordProvider>();
         services.AddSingleton<IWalletStore, WalletStore>();
